+++ packages/ui/src/atoms/Icon/IconSharedProps.ts
@@ -1,0 +1,6 @@
+import { CvElement } from '../../types/CvElement';
+
+export interface IconSharedProps extends CvElement {
+  name: string;
+}
+