+++ packages/ui/src/types/CvElement.ts
@@ -1,0 +1,5 @@
+export interface CvElement {
+  'data-qa'?: string;
+  id?: string;
+}
+